refactoring de la connexion et de l'inscription ainsi que du DAO (mais il reste des erreurs + servlets à adapter)
