+++ src/java/dao/DAO.java
@@ -10,7 +10,7 @@ import java.sql.Statement;
  *
  * @author pierant
  */
-public class DAOImpl {
+public class DAO {
     public static PreparedStatement initialisationRequetePreparee( Connection connexion, String sql, boolean returnGeneratedKeys, Object... objets ) throws SQLException {
         PreparedStatement preparedStatement = connexion.prepareStatement( sql, returnGeneratedKeys ? Statement.RETURN_GENERATED_KEYS : Statement.NO_GENERATED_KEYS );
