--- src/java/dao/DislikeActivityDao.java
@@ -1,27 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package dao;
-
-import beans.Activity;
-import beans.DislikeActivity;
-import java.util.ArrayList;
-import java.util.Map;
-
-/**
- *
- * @author pierant
- */
-public interface DislikeActivityDao {
-    DislikeActivity create(DislikeActivity post) throws DAOException;
-    
-    ArrayList<DislikeActivity> getAll() throws DAOException;
-    DislikeActivity get(int id) throws DAOException;
-    DislikeActivity get(int id_human, int id_post) throws DAOException;
-    ArrayList<DislikeActivity> getByHuman(int id_human) throws DAOException;
-    Map<Integer, DislikeActivity> getHashByHuman(int id_human) throws DAOException;
-    
-    void delete(int id) throws DAOException;
-}
