+++ src/java/dao/ReactionActivityDao.java
@@ -5,17 +5,16 @@
  */
 package dao;
 
-import beans.Activity;
+import beans.DislikeActivity;
 import beans.Reaction;
-import beans.LikeActivity;
-import static dao.DAOImpl.fermeturesSilencieuses;
-import static dao.DAOImpl.initialisationRequetePreparee;
+import beans.ReactionActivity;
+import static dao.DAO.fermeturesSilencieuses;
+import static dao.DAO.initialisationRequetePreparee;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.sql.Timestamp;
 import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -25,27 +24,27 @@ import java.util.Map;
  *
  * @author pierant
  */
-public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao {
-    private DAOFactory daoFactory;
+public class ReactionActivityDao extends BasicDao {
+    private final DAOFactory daoFactory;
     
-    public LikeActivityDaoImpl(DAOFactory daoFactory){
+    public ReactionActivityDao(DAOFactory daoFactory){
         this.daoFactory = daoFactory;
     }
     
-    private static LikeActivity map( ResultSet resultSet ) throws SQLException {
-        LikeActivity activity = new LikeActivity();
+    private static ReactionActivity map( ResultSet resultSet ) throws SQLException {
+        ReactionActivity activity = new ReactionActivity();
         activity.setId(resultSet.getInt( "id"));
         activity.setDate(resultSet.getTimestamp("date").toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime());
         activity.setId_human(resultSet.getInt("id_human"));
-        activity.setReaction(Reaction.LIKE);
+        activity.setReaction(Reaction.valueOf(resultSet.getString( "reaction")));
         activity.setId_post(resultSet.getInt("id_post"));
         
         return activity;
     }
     
-    private static final String SQL_INSERT = "INSERT INTO reactionactivity (id, reaction, id_post) VALUES (?, ?, ?)";
-    @Override
-    public LikeActivity create(LikeActivity activity) throws DAOException {
+    private static final String SQL_INSERT = "INSERT INTO reactionactivity (id, reaction, id_post) VALUES (?, ?::reaction, ?)";
+    
+    public DislikeActivity create(DislikeActivity activity) throws DAOException {
         Connection connexion = null;
         PreparedStatement preparedStatement = null;
         ResultSet valeursAutoGenerees = null;
@@ -72,15 +71,15 @@ public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao
         }
         return activity;
     }
-
-    private static final String SQL_SELECT_ALL = "SELECT a.id as id, date, id_human, id_post FROM reactionactivity r INNER JOIN activity a ON r.id = a.id WHERE r.reaction = 'like'";
-    @Override
-    public ArrayList<LikeActivity> getAll() throws DAOException {
+    
+    private static final String SQL_SELECT_ALL = "SELECT a.id as id, date, id_human, id_post FROM reactionactivity r INNER JOIN activity a ON r.id = a.id";
+    
+    public ArrayList<DislikeActivity> getAll() throws DAOException {
         Connection connexion = null;
         PreparedStatement preparedStatement = null;
         ResultSet result = null;
-        LikeActivity activity = null;
-        ArrayList<LikeActivity> activities = new ArrayList<>();
+        DislikeActivity activity = null;
+        ArrayList<DislikeActivity> activities = new ArrayList<>();
 
         try {
             connexion = daoFactory.getConnection();
@@ -98,14 +97,14 @@ public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao
 
         return activities;
     }
-
+    
     private static final String SQL_SELECT_BY_ID = "SELECT a.id as id, date, id_human, id_post FROM reactionactivity r INNER JOIN activity a ON r.id = a.id WHERE a.id = ?";
-    @Override
-    public LikeActivity get(int id) throws DAOException {
+    
+    public DislikeActivity get(int id) throws DAOException {
         Connection connexion = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        LikeActivity activity = null;
+        DislikeActivity activity = null;
 
         try {
             connexion = daoFactory.getConnection();
@@ -123,13 +122,13 @@ public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao
         return activity;
     }
     
-    private static final String SQL_SELECT_BY_HUMAN_AND_POST = "SELECT a.id as id, date, id_human, id_post FROM reactionactivity r INNER JOIN activity a ON r.id = a.id WHERE id_human = ? AND id_post = ? AND reaction = 'like'";
-    @Override
-    public LikeActivity get(int id_human, int id_post) throws DAOException {
+    private static final String SQL_SELECT_BY_HUMAN_AND_POST = "SELECT a.id as id, date, id_human, id_post FROM reactionactivity r INNER JOIN activity a ON r.id = a.id WHERE id_human = ? AND id_post = ? AND reaction = 'dislike'";
+    
+    public DislikeActivity get(int id_human, int id_post) throws DAOException {
         Connection connexion = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        LikeActivity activity = null;
+        DislikeActivity activity = null;
 
         try {
             connexion = daoFactory.getConnection();
@@ -147,14 +146,14 @@ public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao
         return activity;
     }
     
-    private static final String SQL_SELECT_BY_ID_HUMAN = "SELECT a.id as id, date, id_human, id_post FROM reactionactivity r INNER JOIN activity a ON r.id = a.id WHERE id_human = ? AND reaction = 'like'";
-    @Override
-    public ArrayList<LikeActivity> getByHuman(int id_human) throws DAOException {
+    private static final String SQL_SELECT_BY_ID_HUMAN = "SELECT a.id as id, date, id_human, id_post FROM reactionactivity r INNER JOIN activity a ON r.id = a.id WHERE id_human = ? AND reaction = 'dislike'";
+    
+    public ArrayList<DislikeActivity> getByHuman(int id_human) throws DAOException {
         Connection connexion = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        LikeActivity activity = null;
-        ArrayList<LikeActivity> activities = new ArrayList<>();
+        DislikeActivity activity = null;
+        ArrayList<DislikeActivity> activities = new ArrayList<>();
 
         try {
             connexion = daoFactory.getConnection();
@@ -173,14 +172,14 @@ public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao
 
         return activities;
     }
-   
-    @Override
-    public Map<Integer, LikeActivity> getHashByHuman(int id_human) throws DAOException {
+    
+    
+    public Map<Integer, DislikeActivity> getHashByHuman(int id_human) throws DAOException {
         Connection connexion = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        LikeActivity activity = null;
-        Map<Integer, LikeActivity> activities = new HashMap<>();
+        DislikeActivity activity = null;
+        Map<Integer, DislikeActivity> activities = new HashMap<>();
 
         try {
             connexion = daoFactory.getConnection();
@@ -190,6 +189,7 @@ public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao
                 activity = map( resultSet );
                 activities.put(activity.getId(), activity);
             }
+            
         } catch ( SQLException e ) {
             throw new DAOException( e );
         } finally {
@@ -200,7 +200,7 @@ public class LikeActivityDaoImpl extends BasicDaoImpl implements LikeActivityDao
     }
 
     private static final String SQL_DELETE= "DELETE FROM reactionactivity WHERE id = ?";
-    @Override
+    
     public void delete(int id) throws DAOException {
         super.delete(daoFactory, id, "DELETE FROM activity WHERE id = ?");
         super.delete(daoFactory, id, SQL_DELETE);
