--- src/java/forms/FormValidationException.java
@@ -1,10 +1,0 @@
-package forms;
-
-public class FormValidationException extends Exception {
-    /*
-     * Constructeur
-     */
-    public FormValidationException( String message ) {
-        super( message );
-    }
-}
